+++ src/routes/authRoutes.ts
@@ -1,13 +1,13 @@
 import { Router } from 'express'
-import { handleRegister, handleLogin } from '../controllers/userController'
-import validateLogin  from '../middlewares/validateLogin'
-import validateRegister from '../middlewares/validateRegister'
 
+import { handleLogin, handleRegister } from '../controllers/userController'
+import validateLogin from '../middlewares/validateLogin'
+import validateRegister from '../middlewares/validateRegister'
+import handleValidationErrors from '../middlewares/validationErrorMiddleware'
 
 const router = Router()
 
-router.post('/register', validateRegister, handleRegister)
-
-router.post('/login', validateLogin, handleLogin)
+router.post('/register', validateRegister, handleValidationErrors, handleRegister)
+router.post('/login', validateLogin, handleValidationErrors, handleLogin)
 
 export default router
