+++ src/routes/userRoutes.ts
@@ -1,11 +1,13 @@
-import { Router } from 'express';
-import { getUser, updateUser, deleteUser } from '../controllers/userController';
+import { Router } from 'express'
+
+import { deleteUser, getUser, updateUser } from '../controllers/userController'
 import { validateUpdateUser } from '../middlewares/validateUpdateUser'
+import handleValidationErrors from '../middlewares/validationErrorMiddleware'
 
-const router = Router();
+const router = Router()
 
-router.get('/profile/:id', getUser);
-router.patch('/profile/:id', validateUpdateUser, updateUser);
-router.delete('/profile/:id', deleteUser);
+router.get('/profile/:id', getUser)
+router.patch('/profile/:id', validateUpdateUser, handleValidationErrors, updateUser)
+router.delete('/profile/:id', handleValidationErrors, deleteUser)
 
-export default router;
+export default router
